oon_utilities/pyproject.toml
[tool.poetry]
name = "oon-utilities"
version = "0.1.0"
description = ""
authors = ["christianbonnell <christian.bonnell@gmail.com>"]
readme = "README.md"
packages = [{include = "oon_utilities"}]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.128.0"
google-cloud-storage = "^2.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
oon_utilities = "oon_utilities.scratchpad:main"